chore: added comment

+++ main.py
@@ -4,7 +4,9 @@ import cv2
 
 #! Click points in this order:
 #! Top Left, Top Right, Bottom Right, Bottom Left
-def order_points(rect: list[list[float]]) -> list[list[int]]:
+
+#* Currently Unused
+def order_points(rect: list[list[float]]) -> list[list[float]]:
     points = rect.copy()
     points.sort(key=lambda x: x[1])
